Update the overall repository.

+++ HorrorTracker.Data/Repositories/Interfaces/IOverallRepository.cs
@@ -9,12 +9,12 @@
         /// Retrieves the overall time of the items in the database.
         /// </summary>
         /// <returns>The overall time.</returns>
-        double GetOverallTime();
+        decimal GetOverallTime();
 
         /// <summary>
         /// Retrieves the overall time left of the items not watched in the database.
         /// </summary>
         /// <returns>The overall time left.</returns>
-        double GetOverallTimeLeft();
+        decimal GetOverallTimeLeft();
     }
 }
